requirements.txt
anthropic>=0.86.0
supabase>=2.28.3
httpx>=0.28.0
beautifulsoup4>=4.14.0
feedparser>=6.0.0
python-dotenv>=1.2.0
